sonar issues + color picker refactor
